[feat] 내서재 뷰 레이아웃 구현
