--- app/src/main/java/com/millie/millieshelf/Best/AuthService.kt
@@ -1,10 +1,0 @@
-package com.millie.millieshelf.Best
-
-import com.millie.millieshelf.Best.Data.ResponseBookDto
-import retrofit2.Call
-import retrofit2.http.GET
-
-interface AuthService {
-    @GET("/api/books")
-    fun getBookList(): Call<ResponseBookDto>
-}
