--- app/src/main/java/com/millie/millieshelf/Best/LibService.kt
@@ -1,10 +1,0 @@
-package com.millie.millieshelf.Best
-
-import com.millie.millieshelf.Library.Data.ResponseLibraryDto
-import retrofit2.Call
-import retrofit2.http.GET
-
-interface LibService {
-    @GET("/api/me/collections")
-    fun getBookList(): Call<ResponseLibraryDto>
-}
