--- app/src/main/java/com/millie/millieshelf/Library/Data/BookData.kt
@@ -1,12 +1,0 @@
-package com.millie.millieshelf.Library.Data
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class BookData(
-    @SerialName("id")
-    val id: Long,
-    @SerialName("thumbnail")
-    val thumbnail: String
-)
