--- app/src/main/java/com/millie/millieshelf/Library/Data/ResponseLibData.kt
@@ -1,10 +1,0 @@
-package com.millie.millieshelf.Library.Data
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseLibData(
-    @SerialName("data")
-    val data: List<LibraryCategoryData>
-)
