+++ app/src/main/java/com/millie/millieshelf/data/model/BookCollection.kt
@@ -5,16 +5,16 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BookCollections(
-    @SerialName("books")
-    val books: List<BookCollection> = listOf(),
     @SerialName("category")
     val category: String = "",
+    @SerialName("books")
+    val books: List<BookCollection> = listOf(),
 ) {
     @Serializable
     data class BookCollection(
         @SerialName("id")
-        val id: String = "",
-        @SerialName("imageUrl")
-        val imageUrl: String = "",
+        val id: Long,
+        @SerialName("thumbnail")
+        val thumbnail: String = "",
     )
 }
