+++ app/src/main/res/layout/activity_best.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Best.view.BestActivity">
+    tools:context=".presentation.best.BestActivity">
 
     <ImageView
         android:id="@+id/iv_back"
