+++ app/src/main/res/layout/fragment_mylibrary.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Library.Mylibrary">
+    tools:context=".presentation.library.MylibraryFragment">
 
     <ImageView
         android:id="@+id/iv_background"
